move util to testutil
